Remove dbus-nox11
It made problems, see #956 and #952. Originally this was added, so
our plasma mobile packaging - which targets wayland - doesn't need
to depend on X11 stuff. And because dbus-x11 caused a 1 minute
wait for an X server (#377).

We have yet to figure out if this problem returns, but even if that
happens, it's better than having it completely broken. And we can
probably figure out a better way to fix it (for real).