main/linux-postmarketos-*-dev: fix kernel-scripts
When cross compiling, the kernel scripts would be in the native arch
again. That was already fixed in !1234, however the mime type of the
binaries has changed, so we couldn't detect them properly anymore
(#1659).

This commit fixes it again, and also adds a checks so this won't
silently fail in the future.

[skip ci] because this would not run through anyway (building these
packages takes too long). Fixes #1659.